Improved on docs + stuff
* Improved grammar and fix typos in docs/Memory.md
* Add docs/Math.md
* Move to gcc 12.1.0
* Add nasm to dep target for Arch
* Add #include recommendations to archive/code-style.txt
* Make mem_pallocation store an offset to the header instead of an
  absolute pointer
* Separate typedef struct
* Move some x86_64 specific headers to arch/
* Separate mem.c to mem/mem.c, mem/pmem.c, mem/vmem.c
* Use intmax_t for math operations instead of int
* Remove hello.txt from host/ and initrd/
* Remove some unused headers

* I promise i wont do mega commits again

+++ kernel/stdinc/math.h
@@ -1,26 +1,29 @@
 #ifndef CSTD_MATH_H
 #define CSTD_MATH_H
 
+#include <stddef.h>
+#include <stdint.h>
+
 /**
  * \brief Calculates \a to the power of \a power.
  * \param base
  * \param power
  * \return \a base to the power of \a power.
  */
-int powi(int base, int power);
+intmax_t powi(intmax_t base, intmax_t power);
 
 /**
  * \brief Calculates 10 to the power of \a power.
  * \param power
  * \return 10 the power of \a power.
  */
-int exp10i(int power);
+intmax_t exp10i(intmax_t power);
 
 /**
  * \brief Calculates the absolute value of \a i
  * \param i
  */
-int abs(int i);
+intmax_t abs(intmax_t i);
 
 #define MAX(a,b) ((a) > (b) ? (a) : (b))
 #define MIN(a,b) ((a) < (b) ? (a) : (b))
